simplify tabs
