+++ assets/pages/paper-subscription-landing/paperSubscriptionLandingPageReducer.js
@@ -7,7 +7,32 @@ import type { CommonState } from 'helpers/page/commonReducer';
 import { type PaperBillingPlan } from 'helpers/subscriptions';
 import { getQueryParameter } from 'helpers/url';
 import { ProductPagePlanFormReducerFor, type State as FormState } from 'components/productPage/productPagePlanForm/productPagePlanFormReducer';
-import { tabsReducer, type TabsState } from './components/tabsReducer';
+import { type Tab } from './components/tabs';
+import { type TabActions } from './paperSubscriptionLandingPageActions';
+
+// ----- Tabs ----- //
+
+type TabsState = {
+  active: Tab
+}
+
+const tabsReducer = (state: TabsState = {
+  active: 'collection',
+}, action: TabActions): TabsState => {
+
+  switch (action.type) {
+
+    case 'SET_TAB':
+      return { ...state, active: action.tab };
+
+    default:
+      return state;
+
+  }
+
+};
+
+// ----- Exports ----- //
 
 export type State = {
   common: CommonState,
